fix: property descriptor edge cases
Using the `drafts` object as a source of property descriptors is bad,
because enumerability is not preserved by the `drafts` object.

Thus, we must use `state.base` when `state.copy` is undefined, which
means we need to ensure the property descriptor has a `writable` value
of true. Otherwise, the descriptor would be lying.