fix: unnecessary draft creation
When a draft wraps another draft, it must be careful to never create a draft when accessing its base state.

Also, when the wrapper draft is modified, it copies the wrapped draft at that point in time, which means you should avoid passing a draft to an async producer.

Fixes #343